Only decompress blocks of 128K at a time.
This has the added benefit of allowing truncated backup files to be still
written to an output file. Decompressing all at once in memory fails on a
truncated backup file with and exception message "zlib.error: Error -5 while
decompressing data: incomplete or truncated stream" and does not allow a
partial tarfile to be written.